pyproject.toml
[project]
name = "talk-cad"
version = "0.1.0"
description = "Generative CAD Copilot"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "build123d",
    "fastapi",
    "uvicorn",
    "openai",
    "python-dotenv",
]

[tool.hatch.build.targets.wheel]
packages = ["."]